pyproject.toml
[project]
name = "lazy-rabbit-helper"
version = "0.1.0"
description = ""
authors = [
    {name = "Walter Fan",email = "walter.fan@zoom.us"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pyyaml (==6.0.2)",
    "pydantic (>=2.11.4,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "requests (==2.32.3)",
    "dotenv (==0.9.9)",
    "httpx (==0.27.2)",
    "pyqt5 (==5.15.11)",
    "instructor (>=1.8.0,<2.0.0)",
    "openai (>=1.77.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "lazy_rabbit_helper", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
